pull out TextResults into their own files
