+++ Peer/Apps/AppBuilder/Help.cs
@@ -1,0 +1,8 @@
+namespace Peer.Apps.AppBuilder;
+
+public class Help : TextResult
+{
+    public Help(string text) : base(text)
+    {
+    }
+}
