+++ Peer/Apps/AppBuilder/TextResult.cs
@@ -1,0 +1,11 @@
+namespace Peer.Apps.AppBuilder;
+
+public abstract class TextResult
+{
+    public string Text { get; }
+
+    protected TextResult(string text)
+    {
+        Text = text;
+    }
+}
