+++ Peer/Apps/AppBuilder/UsageError.cs
@@ -1,0 +1,8 @@
+namespace Peer.Apps.AppBuilder;
+
+public class UsageError : TextResult
+{
+    public UsageError(string text) : base(text)
+    {
+    }
+}
