+++ Peer/Apps/AppBuilder/Version.cs
@@ -1,0 +1,9 @@
+namespace Peer.Apps.AppBuilder;
+
+public class Version : TextResult
+{
+    public Version(string text) : base(text)
+    {
+        
+    }
+}
